requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
pydantic==2.3.0
psycopg2-binary==2.9.7
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
openai==0.27.8
pandas==2.1.0
pytest==7.4.0
httpx==0.24.1
alembic==1.12.0